pyproject.toml
[project]
name = "fs-security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Alejandro Ouslan" },
  { email = "alejandro.ouslan@upr.edu" },
]
license = { file = "LICENSE" }
dependencies = [
  "alembic>=1.14.0",
  "geoalchemy2>=0.17.0",
  "geopandas>=1.0.1",
  "ibis-framework[sqlite]>=9.5.0",
  "polars>=1.19.0",
  "psycopg2-binary>=2.9.10",
  "requests>=2.32.3",
  "sqlmodel>=0.0.22",
  "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["jupyter>=1.1.1", "pynvim>=0.5.2"]
